Modification of the encoding and iteration.
Encoding is now more compact (a decent size of non-accessible boards are
left out). This is at the cost of a more expensive encoding function but
this is counter-balanced by an optimization of the iterator: it now
yields the awari board and it's code, diminishing the number of calls to
`Awari::encode`.